requirements.txt
notebook==7.2.0
jupyter==1.0.0
ipykernel==6.29.4 
matplotlib==3.9.0
numpy==1.26.4
scikit-learn==1.5.0
instructor==1.2.6
openai==1.27.0
datasets==2.19.1
loguru==0.7.2
lancedb==0.6.13
pydantic-settings==2.2.1
tantivy==0.22.0
cohere==5.5.6